OneDrive/Documents/trading_ai/requirements.txt
oandapyV20
requests
pytz
python-dotenv
feedparser
flask
